src/List/Divider/Divider.component.spec.tsx
// REACT
import * as React from 'react';
import * as renderer from 'react-test-renderer';
// VENDOR
import { shallow, mount } from 'enzyme';
// COMPONENT
import { Divider } from './Divider.component';

describe('Component: Divider', () => {
    it('should be defined', () => {
        const subject = <Divider />;
        const wrapper = mount(subject);
        const component = shallow(subject);

        /* tslint:disable */
        console.log(subject);
        console.log(wrapper.debug());
        console.log(component.debug());

        /* tslint:enable */
        expect(subject).toBeDefined();
        expect(wrapper).toBeDefined();
        expect(component).toBeDefined();
        const tree = renderer.create(subject).toJSON();

        expect(tree).toMatchSnapshot();
    });
});
